// backend/src/generated-api/models/pick-options.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Survivor Fantasy League - OpenAPI 3.0.3
 * This is the Survivor Fantasy League Spec for API's
 *
 * OpenAPI spec version: 1.0.11
 * Contact: pklein111697@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { BinaryPickOptions } from './binary-pick-options';
import { ColorPickOptions } from './color-pick-options';
import { PickOptionTypeEnum } from './pick-option-type-enum';
import { SurvivorPickOptions } from './survivor-pick-options';
import { TribePickOptions } from './tribe-pick-options';
 /**
 * 
 *
 * @export
 * @interface PickOptions
 */
export interface PickOptions {

    /**
     * @type {PickOptionTypeEnum}
     * @memberof PickOptions
     */
    pickOptionType: PickOptionTypeEnum;

    /**
     * The minimum number of selections required
     *
     * @type {number}
     * @memberof PickOptions
     * @example 1
     */
    minNumSelections: number;

    /**
     * The maximum number of selections allowed
     *
     * @type {number}
     * @memberof PickOptions
     * @example 1
     */
    maxNumSelections: number;

    /**
     * Whether the pick option allows for a \"none\" option
     *
     * @type {boolean}
     * @memberof PickOptions
     * @example false
     */
    noneOptionAllowed: boolean;

    /**
     * @type {SurvivorPickOptions | ColorPickOptions | TribePickOptions | BinaryPickOptions}
     * @memberof PickOptions
     */
    options: SurvivorPickOptions | ColorPickOptions | TribePickOptions | BinaryPickOptions;
}